# docs/schema/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "Anomaly" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "$ref" : "#/components/schemas/AnomalyType"
          },
          "standId" : {
            "type" : "string"
          },
          "standNom" : {
            "type" : "string"
          },
          "date" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "AnomalyType" : {
        "type" : "string",
        "enum" : [ "STAND_JAMAIS_OUVERT", "FENETRE_SANS_EFFET", "SEGMENT_TROP_COURT" ]
      },
...
      "Avertissement" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "$ref" : "#/components/schemas/TypeAvertissement"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "CauseInfaisabilite" : {
        "type" : "object",
        "required" : [ "capacite", "demande", "manque" ],
        "properties" : {
          "type" : {
            "$ref" : "#/components/schemas/TypeCauseInfaisabilite"
          },
          "severite" : {
            "$ref" : "#/components/schemas/SeveriteInfaisabilite"
          },
          "message" : {
            "type" : "string"
          },
          "creneauId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "date" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "heureDebut" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "heureFin" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "standIds" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "contrainteIds" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "demande" : {
            "type" : "integer",
            "format" : "int32"
          },
          "capacite" : {
            "type" : "integer",
            "format" : "int32"
          },
          "manque" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "Coupure" : {
        "required" : [ "nombreStands" ],
        "type" : "object",
        "properties" : {
          "date" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "heure" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "standIds" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "nombreStands" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "Creneau" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "jour" : {
            "type" : "integer",
            "format" : "int32"
          },
          "date" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "heureDebut" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "heureFin" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "famille" : {
            "type" : "integer",
            "format" : "int32"
          },
          "couverturePause" : {
            "type" : "boolean"
          },
          "modifieLe" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "dureeMinutes" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "DayOfWeek" : {
        "type" : "string",
        "enum" : [ "MONDAY", "TUESDAY", "WEDNESDAY", "THURSDAY", "FRIDAY", "SATURDAY", "SUNDAY" ]
      },
...
      "DerivationRequest" : {
        "required" : [ "remplacer" ],
        "type" : "object",
        "properties" : {
          "dateDebut" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "dateFin" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "heureFermeture" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "dureeMinimaleMinutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "remplacer" : {
            "type" : "boolean"
          }
        }
      },
...
      "FeasibilityReport" : {
        "type" : "object",
        "required" : [ "feasible", "manqueAnimateurs", "totalCauses" ],
        "properties" : {
          "feasible" : {
            "type" : "boolean"
          },
          "manqueAnimateurs" : {
            "type" : "integer",
            "format" : "int32"
          },
          "causes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CauseInfaisabilite"
            }
          },
          "totalCauses" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "FenetreHoraire" : {
        "type" : "object",
        "properties" : {
          "heureDebut" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "heureFin" : {
            "$ref" : "#/components/schemas/LocalTime"
          },
          "effectif" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "GridAnomaly" : {
        "type" : "object",
        "properties" : {
          "severite" : {
            "$ref" : "#/components/schemas/SeveriteGrille"
          },
          "type" : {
            "$ref" : "#/components/schemas/GridAnomalyType"
          },
          "date" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "GridAnomalyType" : {
        "type" : "string",
        "enum" : [ "CRENEAU_INCOMPLET", "DUREE_NULLE", "DOUBLON", "REPOS_QUOTIDIEN_IMPOSSIBLE", "CHEVAUCHEMENT", "TROU_DANS_LA_JOURNEE", "AMPLITUDE_PLUS_COURTE_QUE_LA_VACATION_MINIMALE", "VACATION_TROP_LONGUE", "DATE_ISOLEE" ]
      },
...
      "Instant" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T16:15:50Z" ]
      },
...
      "LocalDate" : {
        "type" : "string",
        "format" : "date",
        "examples" : [ "2022-03-10" ]
      },
...
      "LocalTime" : {
        "type" : "string",
        "format" : "local-time",
        "examples" : [ "13:45:30.123456789" ],
        "externalDocs" : {
          "description" : "As defined by 'partial-time' in RFC3339",
          "url" : "https://www.rfc-editor.org/rfc/rfc3339.html#section-5.6"
        }
      },
...
      "ModeGrilleCreneaux" : {
        "type" : "string",
        "enum" : [ "AMPLITUDES", "VACATIONS" ]
      },
...
      "RapportDerivation" : {
        "type" : "object",
        "properties" : {
          "nombreGeneres" : {
            "type" : "integer",
            "format" : "int32"
          },
          "creneaux" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Creneau"
            }
          },
          "coupures" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Coupure"
            }
          },
          "joursSansFenetre" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "date",
              "examples" : [ "2022-03-10" ]
            }
          },
          "controle" : {
            "$ref" : "#/components/schemas/RapportGrille"
          }
        }
      },
...
      "RapportGrille" : {
        "type" : "object",
        "required" : [ "nombreCreneaux" ],
        "properties" : {
          "mode" : {
            "$ref" : "#/components/schemas/ModeGrilleCreneaux"
          },
          "nombreCreneaux" : {
            "type" : "integer",
            "format" : "int32"
          },
          "anomalies" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/GridAnomaly"
            }
          },
          "ouvertures" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Anomaly"
            }
          },
          "faisabilite" : {
            "$ref" : "#/components/schemas/FeasibilityReport"
          }
        }
      },
...
      "RapportRecurrence" : {
        "required" : [ "nombreGeneres" ],
        "type" : "object",
        "properties" : {
          "nombreGeneres" : {
            "type" : "integer",
            "format" : "int32"
          },
          "creneaux" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Creneau"
            }
          },
          "controle" : {
            "$ref" : "#/components/schemas/RapportGrille"
          }
        }
      },
...
      "RecurrenceRequest" : {
        "type" : "object",
        "properties" : {
          "jours" : {
            "$ref" : "#/components/schemas/TypeJoursHoraire"
          },
          "dateDebut" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "dateFin" : {
            "$ref" : "#/components/schemas/LocalDate"
          },
          "joursSemaine" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "$ref" : "#/components/schemas/DayOfWeek"
            }
          },
          "dates" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "type" : "string",
              "format" : "date",
              "examples" : [ "2022-03-10" ]
            }
          },
          "exclusions" : {
            "type" : "array",
            "uniqueItems" : true,
            "items" : {
              "type" : "string",
              "format" : "date",
              "examples" : [ "2022-03-10" ]
            }
          },
          "fenetres" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/FenetreHoraire"
            }
          }
        }
      },
      "ReferenceUsage" : {
        "required" : [ "affectations", "contraintesAdHoc", "verrouillages" ],
        "type" : "object",
        "properties" : {
          "affectations" : {
            "type" : "integer",
            "format" : "int32"
          },
          "contraintesAdHoc" : {
            "type" : "integer",
            "format" : "int32"
          },
          "verrouillages" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "SeveriteGrille" : {
        "type" : "string",
        "enum" : [ "ERREUR", "AVERTISSEMENT" ]
      },
      "SeveriteInfaisabilite" : {
        "type" : "string",
        "enum" : [ "CRITIQUE", "ELEVE" ]
      },
...
      "TypeAvertissement" : {
        "type" : "string",
        "enum" : [ "INDISPONIBILITE_HORS_EVENEMENT", "INDISPONIBILITE_JOUR_SANS_CRENEAU", "MINEUR_PENDANT_EVENEMENT", "CRENEAU_HORS_OUVERTURE_STANDS", "CRENEAU_DEBORDE_OUVERTURE_STANDS", "STAND_EXCEPTION_HORS_EVENEMENT", "STAND_FENETRE_SANS_EFFET", "STAND_JAMAIS_OUVERT" ]
      },
      "TypeCauseInfaisabilite" : {
        "type" : "string",
        "enum" : [ "CRENEAU_SOUS_EFFECTIF", "CONTRAINTES_AD_HOC_CONTRADICTOIRES" ]
      },
...
      "TypeJoursHoraire" : {
        "type" : "string",
        "enum" : [ "TOUS", "JOURS_SEMAINE", "PLAGE", "DATES" ]
      },
...
      "WrittenCreneau" : {
        "type" : "object",
        "properties" : {
          "creneau" : {
            "$ref" : "#/components/schemas/Creneau"
          },
          "avertissements" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Avertissement"
            }
          }
        }
      },
...
    }
  },
  "paths" : {
...
    "/api/creneaux" : {
      "get" : {
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Creneau"
                  }
                }
              }
            }
          }
        },
        "summary" : "List Creneaux",
        "tags" : [ "Creneau Resource" ]
      },
      "post" : {
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Creneau"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WrittenCreneau"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Create Creneau",
        "tags" : [ "Creneau Resource" ]
      }
    },
    "/api/creneaux/controle" : {
      "get" : {
        "parameters" : [ {
          "name" : "mode",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RapportGrille"
                }
              }
            }
          }
        },
        "summary" : "Validate Grid",
        "tags" : [ "Creneau Resource" ]
      }
    },
    "/api/creneaux/derivation" : {
      "post" : {
        "parameters" : [ {
          "name" : "mode",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DerivationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RapportDerivation"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Apply Derivation",
        "tags" : [ "Creneau Resource" ]
      }
    },
    "/api/creneaux/derivation/apercu" : {
      "post" : {
        "parameters" : [ {
          "name" : "mode",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DerivationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RapportDerivation"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Preview Derivation",
        "tags" : [ "Creneau Resource" ]
      }
    },
...
    "/api/creneaux/recurrence" : {
      "post" : {
        "parameters" : [ {
          "name" : "mode",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RecurrenceRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RapportRecurrence"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Create Recurrence",
        "tags" : [ "Creneau Resource" ]
      }
    },
    "/api/creneaux/recurrence/apercu" : {
      "post" : {
        "parameters" : [ {
          "name" : "mode",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RecurrenceRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RapportRecurrence"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Preview Recurrence",
        "tags" : [ "Creneau Resource" ]
      }
    },
    "/api/creneaux/usages" : {
      "get" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReferenceUsage"
                }
              }
            }
          }
        },
        "summary" : "Count Creneau Usages",
        "tags" : [ "Creneau Resource" ]
      }
    },
    "/api/creneaux/{id}" : {
      "put" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Creneau"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WrittenCreneau"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "summary" : "Update Creneau",
        "tags" : [ "Creneau Resource" ]
      },
      "delete" : {
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : { }
              }
            }
          }
        },
        "summary" : "Delete Creneau",
        "tags" : [ "Creneau Resource" ]
      }
    },
...
  },
  "info" : {
    "title" : "planning-equipes API",
    "version" : "999-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}